requirements.txt
python-frontmatter==1.1.0
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pandas==2.2.3
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
